Move get demographics to seperate file
