+++ VAMobile/src/screens/HomeScreen/HomeScreen.tsx
@@ -17,7 +17,7 @@ import { logAnalyticsEvent } from 'utils/analytics'
 import { logCOVIDClickAnalytics } from 'store/slices/vaccineSlice'
 import { stringToTitleCase } from 'utils/formattingUtils'
 import { useAppDispatch, useRouteNavigation, useTheme } from 'utils/hooks'
-import { useDemographics } from 'api/demographics'
+import { useDemographics } from 'api/demographics/getDemographics'
 import { useSelector } from 'react-redux'
 import ContactInformationScreen from './ProfileScreen/ContactInformationScreen'
 import ContactVAScreen from './ContactVAScreen/ContactVAScreen'
